Change migration version to 38
