+++ src/ert/storage/local_storage.py
@@ -31,7 +31,7 @@ from .realization_storage_state import RealizationStorageState
 
 logger = logging.getLogger(__name__)
 
-_LOCAL_STORAGE_VERSION = 37
+_LOCAL_STORAGE_VERSION = 38
 
 
 def open_storage(
@@ -546,6 +546,7 @@ class LocalStorage(BaseMode):
             to35,
             to36,
             to37,
+            to38,
         )
 
         try:  # noqa: PLW0717
@@ -612,6 +613,7 @@ class LocalStorage(BaseMode):
                     34: to35,
                     35: to36,
                     36: to37,
+                    37: to38,
                 }
                 for from_version in range(version, _LOCAL_STORAGE_VERSION):
                     migrations[from_version].migrate(self.path)
