+++ tests/ert/unit_tests/storage/migration/test_to37.py
@@ -1,89 +1,90 @@
+import json
 from pathlib import Path
 
-import polars as pl
+import pytest
 
-from ert.storage.blob_data import BlobStorageData, EverestBatchData
 from ert.storage.migration.to37 import migrate
 
 
-def _write_batch_parquet(ens_dir: Path, dataframe_name: str) -> pl.DataFrame:
-    df = pl.DataFrame({"batch_id": [0], dataframe_name: [1.5]})
-    df.write_parquet(ens_dir / f"{dataframe_name}.parquet")
-    return df
+def _write_index(ens_path: Path, index_data: dict) -> Path:
+    ens_path.mkdir(parents=True)
+    index_file = ens_path / "index.json"
+    index_file.write_text(json.dumps(index_data), encoding="utf-8")
+    return index_file
 
 
-def test_that_migration_moves_batch_parquet_files_into_blobs(tmp_path):
+ENSEMBLE_ID = "6ec3d3f0-8b5b-4a3f-9c1a-3f6a3b0f1c11"
+EXPERIMENT_ID = "1f2e3d4c-5b6a-4978-8899-aabbccddeeff"
+PRIOR_ENSEMBLE_ID = "9a8b7c6d-5e4f-4a3b-8c2d-1e0f9a8b7c6d"
+
+
+def _index_data(**overrides) -> dict:
+    data = {
+        "id": ENSEMBLE_ID,
+        "experiment_id": EXPERIMENT_ID,
+        "ensemble_size": 10,
+        "iteration": 3,
+        "name": "batch_0",
+        "prior_ensemble_id": PRIOR_ENSEMBLE_ID,
+        "started_at": "2023-01-01T00:00:00+00:00",
+        "everest_realization_info": {
+            "0": {"model_realization": 0, "perturbation": -1},
+            "1": {"model_realization": 0, "perturbation": 0},
+        },
+    }
+    data.update(overrides)
+    return data
+
+
+@pytest.mark.parametrize(
+    ("original", "expected"),
+    [
+        pytest.param(
+            _index_data(is_improvement=True),
+            _index_data(),
+            id="removes_is_improvement_true",
+        ),
+        pytest.param(
+            _index_data(is_improvement=False),
+            _index_data(),
+            id="removes_is_improvement_false",
+        ),
+        pytest.param(
+            _index_data(is_improvement=None),
+            _index_data(),
+            id="removes_is_improvement_none",
+        ),
+        pytest.param(
+            _index_data(),
+            _index_data(),
+            id="leaves_untouched_when_missing",
+        ),
+    ],
+)
+def test_that_migration_updates_ensemble_index(tmp_path, original, expected):
     root = tmp_path / "project"
-    ens_dir = root / "ensembles" / "ens-1"
-    ens_dir.mkdir(parents=True)
-
-    objectives = _write_batch_parquet(ens_dir, "batch_objectives")
-    gradient = _write_batch_parquet(ens_dir, "batch_objective_gradient")
+    index_file = _write_index(root / "ensembles" / "ensemble_1", original)
 
     migrate(root)
 
-    assert not (ens_dir / "batch_objectives.parquet").exists()
-    assert not (ens_dir / "batch_objective_gradient.parquet").exists()
-
-    blob_dir = ens_dir / "blobs"
-    metas = [
-        BlobStorageData.model_validate_json(meta_file.read_text(encoding="utf-8"))
-        for meta_file in sorted(blob_dir.glob("*.blob.json"))
-    ]
-
-    by_name = {meta.blob_info.dataframe_name: meta for meta in metas}
-    assert set(by_name) == {"batch_objectives", "batch_objective_gradient"}
-
-    for dataframe_name, expected in (
-        ("batch_objectives", objectives),
-        ("batch_objective_gradient", gradient),
-    ):
-        meta = by_name[dataframe_name]
-        assert isinstance(meta.blob_info, EverestBatchData)
-        assert meta.file_type == "application/parquet"
-        assert meta.name == dataframe_name
-        blob_bytes = (blob_dir / meta.uri).read_bytes()
-        assert meta.file_size == len(blob_bytes)
-        assert pl.read_parquet(blob_dir / meta.uri).equals(expected)
+    assert json.loads(index_file.read_text(encoding="utf-8")) == expected
 
 
-def test_that_migration_leaves_non_batch_parquet_files_untouched(tmp_path):
-    root = tmp_path / "project"
-    ens_dir = root / "ensembles" / "ens-1"
-    realization_dir = ens_dir / "realization-0"
-    realization_dir.mkdir(parents=True)
-
-    responses = pl.DataFrame({"response_key": ["summary"], "values": [1.0]})
-    responses.write_parquet(realization_dir / "responses.parquet")
-
-    migrate(root)
-
-    assert (realization_dir / "responses.parquet").exists()
-    assert not (ens_dir / "blobs").exists()
-
-
-def test_that_migration_is_a_noop_without_ensembles_directory(tmp_path):
+def test_that_migration_does_not_fail_on_unexpectedly_structured_dirs(tmp_path):
     root = tmp_path / "project"
     root.mkdir()
-
     migrate(root)
 
-    assert list(root.iterdir()) == []
-
-
-def test_that_migration_records_only_present_batch_dataframes_as_blobs(tmp_path):
-    root = tmp_path / "project"
-    ens_dir = root / "ensembles" / "ens-1"
-    ens_dir.mkdir(parents=True)
-
-    _write_batch_parquet(ens_dir, "batch_constraints")
-
+    ensembles_dir = root / "ensembles"
+    ensembles_dir.mkdir()
     migrate(root)
 
-    blob_metas = list((ens_dir / "blobs").glob("*.blob.json"))
-    assert len(blob_metas) == 1
-    meta = BlobStorageData.model_validate_json(
-        blob_metas[0].read_text(encoding="utf-8")
-    )
-    assert isinstance(meta.blob_info, EverestBatchData)
-    assert meta.blob_info.dataframe_name == "batch_constraints"
+    not_an_ensemble = ensembles_dir / "not_a_directory.json"
+    not_an_ensemble.write_text("{}", encoding="utf-8")
+    migrate(root)
+
+    ensemble_without_index_json = ensembles_dir / "ensemble_1"
+    ensemble_without_index_json.mkdir()
+    migrate(root)
+
+    assert not_an_ensemble.read_text(encoding="utf-8") == "{}"
